requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
Flask-SQLAlchemy==2.1
greenlet==1.1.1
SQLAlchemy==1.4.23
boto3==1.18.11
requests==2.26.0
